Changes I made 4/22/2017
I changed quite a bit of the schema because there were many bad FK
relationships that made no sense compared to the UML we had. I believe
I've fixed all of these errors. I included a Java script to populate the
zipCode table with all the proper data. The CSV with the zip code info
is also here. Finally, I included the proper SQL to create all tables
and insert all data to the database.