Swapped some tables
I removed the original Sprint Backlog table because its relationships didn't make sense. Instead, I made the UserSprintStory table into the Sprint Backlog table. The UserSprintStory existed because we said that there was a many-to-many relationship between Sprint and User Story. I believe it's only one-to-many. If a user story is not completed in the current sprint, its sprintNum attribute can simply be updated in the Sprint Backlog. This design is much easier to work with and makes more sense.