All of the code I did today
Pretty much all we need to do is glue it together. I've added all of the menu options (features that he wants us to have). The majority of the options are covered by the functions I've already written. Some are not, though. There's also a disgusting SQL query for listing all developers that are part of a sprint. It technically works, but I'm sure it'll either break, or won't work on some specific dataset. I really don't know.